refactor: absorb `wdl-cli` into `sprocket`
This is a discrete step on the path to more extensive refactoring. It's likely that some of the
logic this PR moves from `wdl-cli` into `sprocket` will actually end up in `wdl-engine` eventually,
but this first step is meant to be as mechanical as possible to simplify review despite the amount
of code motion.

Beyond moving and renaming, there are three primary differences in the resulting code:

- Diagnostics-related code that was previously in `src/lib.rs` now has its own module.

- `src/main.rs` is now a minimal wrapper around a single entrypoint `sprocket_main`, now exported
  from the `sprocket` library crate.

- The `pub` surface area of the `sprocket` library crate has been minimized down to those
  definitions necessary to support the `sprocket` binary and the `cli` integration test. This
  uncovered a handful of dead methods which have been removed.
