+++ Cargo.toml
@@ -5,7 +5,6 @@ members = [
     "crates/wdl",
     "crates/wdl-analysis",
     "crates/wdl-ast",
-    "crates/wdl-cli",
     "crates/wdl-doc",
     "crates/wdl-engine",
     "crates/wdl-format",
@@ -110,7 +109,7 @@ url = "2.5.4"
 urlencoding = "2.1.3"
 uuid = "1.18.0"
 walkdir = "2.5.0"
-wdl = { version = "0.19.0", path = "crates/wdl", features = ["cli", "codespan", "engine", "lsp", "doc"] }
+wdl = { version = "0.19.0", path = "crates/wdl", features = ["codespan", "engine", "lint", "lsp", "doc"] }
 which = "8.0.0"
 
 [workspace.lints.rust]
@@ -163,13 +162,17 @@ indexmap.workspace = true
 indicatif.workspace = true
 nonempty.workspace = true
 opener.workspace = true
+path-clean.workspace = true
 pest.workspace = true
 pretty_assertions.workspace = true
+regex.workspace = true
+reqwest.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 serde_yaml_ng.workspace = true
 similar.workspace = true
 strum.workspace = true
+thiserror.workspace = true
 tokio.workspace = true
 tokio-util.workspace = true
 toml.workspace = true
