+++ crates/RELEASE.md
@@ -18,7 +18,7 @@ Once the bump PR merges, tag the HEAD commit with `wdl-v{VERSION}` (where `VERSI
 
 Next up is making a GitHub release, which should be done manually. Please review the most recent releases, as we sometimes change the GitHub "Release Notes" formatting (we're still doing some experimentation here). What follows is what I (a-frantz) have done most recently (`wdl-v0.15.0`), but I would not be shocked if this changes in the future and someone (me) forgets to update this document as the processs is tweaked.
 
-Each crate's most recent CHANGELOG entries should be copy and pasted into the release notes. These should be ordered topologically (starting with `wdl-grammar`, ending with `wdl-cli` or `wdl` itself if that had non-dependency changes). Format each section so that it looks like 
+Each crate's most recent CHANGELOG entries should be copy and pasted into the release notes. These should be ordered topologically (starting with `wdl-grammar`, ending with `wdl` itself if that had non-dependency changes). Format each section so that it looks like 
 ```
 ### `wdl-<crate name>`
