--- crates/wdl-cli/Cargo.toml
@@ -1,42 +1,0 @@
-[package]
-name = "wdl-cli"
-version = "0.7.0"
-description = "Facilities for building command-line tools using the `wdl` crates"
-rust-version.workspace = true
-license.workspace = true
-edition.workspace = true
-authors.workspace = true
-homepage.workspace = true
-repository = "https://github.com/stjude-rust-labs/sprocket/tree/main/crates/wdl-cli"
-documentation = "https://docs.rs/wdl-cli"
-
-[dependencies]
-anyhow.workspace = true
-codespan-reporting.workspace = true
-crankshaft.workspace = true
-futures.workspace = true
-indexmap.workspace = true
-nonempty.workspace = true
-path-clean.workspace = true
-regex.workspace = true
-reqwest.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-serde_yaml_ng.workspace = true
-shellexpand.workspace = true
-thiserror.workspace = true
-tokio.workspace = true
-tokio-util.workspace = true
-tracing.workspace = true
-url.workspace = true
-wdl-analysis = { path = "../wdl-analysis", version = "0.14.0" }
-wdl-ast = { path = "../wdl-ast", version = "0.18.0" }
-wdl-engine = { path = "../wdl-engine", version = "0.9.0" }
-wdl-lint = { path = "../wdl-lint", version = "0.17.0" }
-
-[dev-dependencies]
-pretty_assertions.workspace = true
-tempfile.workspace = true
-
-[lints]
-workspace = true
