+++ crates/wdl-lint/src/rules/trailing_comma.rs
