+++ crates/wdl/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+#### Removed
+
+* Removed the `cli` feature and module ([#450](https://github.com/stjude-rust-labs/sprocket/pull/450)).
+
 ## 0.19.0 - 10-14-2025
 
 #### Dependencies
