+++ crates/wdl/Cargo.toml
@@ -20,7 +20,6 @@ default = [
     "ast",
     "engine",
     "format",
-    "cli",
     "doc",
     "grammar",
     "lint",
@@ -28,7 +27,6 @@ default = [
 ]
 analysis = ["dep:wdl-analysis"]
 ast = ["dep:wdl-ast"]
-cli = ["dep:wdl-cli"]
 doc = ["dep:wdl-doc"]
 engine = ["dep:wdl-engine"]
 format = ["dep:wdl-format"]
@@ -59,7 +57,6 @@ tracing-subscriber = { workspace = true, optional = true }
 url = { workspace = true, optional = true }
 wdl-analysis = { path = "../wdl-analysis", version = "0.14.0", optional = true }
 wdl-ast = { path = "../wdl-ast", version = "0.18.0", optional = true }
-wdl-cli = { path = "../wdl-cli", version = "0.7.0", optional = true }
 wdl-doc = { path = "../wdl-doc", version = "0.9.0", optional = true }
 wdl-engine = { path = "../wdl-engine", version = "0.9.0", optional = true }
 wdl-format = { path = "../wdl-format", version = "0.12.0", optional = true }
