+++ src/analysis.rs
@@ -8,19 +8,19 @@ use futures::future::BoxFuture;
 use nonempty::NonEmpty;
 use tracing::info;
 use tracing::warn;
-use wdl_analysis::Analyzer;
-use wdl_analysis::DiagnosticsConfig;
-use wdl_analysis::ProgressKind;
-use wdl_analysis::Validator;
-use wdl_lint::Linter;
+use wdl::analysis::Analyzer;
+use wdl::analysis::DiagnosticsConfig;
+use wdl::analysis::ProgressKind;
+use wdl::analysis::Validator;
+use wdl::lint::Linter;
 
 mod results;
 mod source;
 
 pub use results::AnalysisResults;
 pub use source::Source;
-use wdl_lint::Rule;
-use wdl_lint::TagSet;
+use wdl::lint::Rule;
+use wdl::lint::TagSet;
 
 /// The type of the initialization callback.
 type InitCb = Box<dyn Fn() + 'static>;
@@ -65,12 +65,6 @@ impl Analysis {
     /// Adds multiple sources to the analysis.
     pub fn extend_sources(mut self, source: impl IntoIterator<Item = Source>) -> Self {
         self.sources.extend(source);
-        self
-    }
-
-    /// Adds a rule to the excepted rules list.
-    pub fn add_exception(mut self, rule: impl Into<String>) -> Self {
-        self.exceptions.insert(rule.into());
         self
     }
 
@@ -122,7 +116,7 @@ impl Analysis {
         if self.enabled_lint_tags.count() > 0 && tracing::enabled!(tracing::Level::INFO) {
             let mut enabled_rules = vec![];
             let mut disabled_rules = vec![];
-            for rule in wdl_lint::rules() {
+            for rule in wdl::lint::rules() {
                 if is_rule_enabled(
                     &self.enabled_lint_tags,
                     &self.disabled_lint_tags,
@@ -137,7 +131,7 @@ impl Analysis {
             info!("enabled lint rules: {:?}", enabled_rules);
             info!("disabled lint rules: {:?}", disabled_rules);
         }
-        let config = wdl_analysis::Config::default()
+        let config = wdl::analysis::Config::default()
             .with_diagnostics_config(get_diagnostics_config(&self.exceptions))
             .with_ignore_filename(self.ignore_filename);
 
@@ -195,12 +189,12 @@ impl Default for Analysis {
 
 /// Warns about any unknown rules.
 fn warn_unknown_rules(exceptions: &HashSet<String>) {
-    let mut names = wdl_analysis::rules()
+    let mut names = wdl::analysis::rules()
         .iter()
         .map(|rule| rule.id().to_owned())
         .collect::<Vec<_>>();
 
-    names.extend(wdl_lint::rules().iter().map(|rule| rule.id().to_owned()));
+    names.extend(wdl::lint::rules().iter().map(|rule| rule.id().to_owned()));
 
     let mut unknown = exceptions
         .iter()
@@ -222,7 +216,7 @@ fn warn_unknown_rules(exceptions: &HashSet<String>) {
 /// Gets the rules as a diagnositics configuration with the excepted rules
 /// removed.
 fn get_diagnostics_config(exceptions: &HashSet<String>) -> DiagnosticsConfig {
-    DiagnosticsConfig::new(wdl_analysis::rules().into_iter().filter(|rule| {
+    DiagnosticsConfig::new(wdl::analysis::rules().into_iter().filter(|rule| {
         !exceptions
             .iter()
             .any(|exception| exception.eq_ignore_ascii_case(rule.id()))
@@ -253,7 +247,7 @@ fn get_lint_visitor(
     disabled_lint_tags: &TagSet,
     exceptions: &HashSet<String>,
 ) -> Linter {
-    Linter::new(wdl_lint::rules().into_iter().filter(|rule| {
+    Linter::new(wdl::lint::rules().into_iter().filter(|rule| {
         is_rule_enabled(
             enabled_lint_tags,
             disabled_lint_tags,
