+++ src/commands/check.rs
@@ -13,8 +13,6 @@ use strum::VariantArray;
 use tracing::info;
 use wdl::ast::AstNode;
 use wdl::ast::Severity;
-use wdl::cli::Analysis;
-use wdl::cli::analysis::Source;
 use wdl::lint::Tag;
 use wdl::lint::TagSet;
 use wdl::lint::find_nearest_rule;
@@ -22,9 +20,11 @@ use wdl::lint::find_nearest_rule;
 use super::explain::ALL_RULE_IDS;
 use super::explain::ALL_TAG_NAMES;
 use crate::IGNORE_FILENAME;
-use crate::Mode;
-use crate::emit_diagnostics;
-use crate::get_diagnostics_display_config;
+use crate::analysis::Analysis;
+use crate::analysis::Source;
+use crate::diagnostics::Mode;
+use crate::diagnostics::emit_diagnostics;
+use crate::diagnostics::get_diagnostics_display_config;
 
 /// The [`Tag`]s which will run with the default `lint` configuration.
 const DEFAULT_TAG_SET: TagSet = TagSet::new(&[
