+++ tests/fixtures/inputs_three.yml
